+++ google_analytics/templatetags/google_analytics_tags.py
@@ -1,12 +1,9 @@
-import urllib
-import urlparse
-
 from django import template
 from django.core.urlresolvers import reverse
 from django.conf import settings
-
 from google_analytics import CAMPAIGN_TRACKING_PARAMS
 
+from six.moves.urllib.parse import parse_qs, urlencode, urlparse
 
 register = template.Library()
 
@@ -16,7 +13,7 @@ def google_analytics(context, tracking_code=None, debug=False):
     if not tracking_code:
         try:
             assert settings.GOOGLE_ANALYTICS['google_analytics_id']
-        except:
+        except KeyError:
             return ''
     # attempt get the request from the context
     request = context.get('request', None)
@@ -35,12 +32,12 @@ def google_analytics(context, tracking_code=None, debug=False):
         params['r'] = referer
     # remove collected parameters from the path and pass it on
     path = request.path
-    parsed_url = urlparse.urlparse(path)
-    query = urlparse.parse_qs(parsed_url.query)
+    parsed_url = urlparse(path)
+    query = parse_qs(parsed_url.query)
     for param in params:
         if param in query:
             del query[param]
-    query = urllib.urlencode(query)
+    query = urlencode(query)
     new_url = parsed_url._replace(query=query)
     params['p'] = new_url.geturl()
     params['tracking_code'] = tracking_code or settings.GOOGLE_ANALYTICS[
@@ -51,5 +48,5 @@ def google_analytics(context, tracking_code=None, debug=False):
     # build and return the url
     url = reverse('google-analytics')
     if len(params) > 0:
-        url += '?&' + urllib.urlencode(params)
+        url += '?&' + urlencode(params)
     return url