+++ google_analytics/tests/test_ga.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 import responses
-import pytest
 
 from django.http import HttpResponse
 from django.test import TestCase, override_settings
@@ -10,12 +9,12 @@ from django.test.client import RequestFactory
 from django.contrib.sessions.middleware import SessionMiddleware
 
 from google_analytics.utils import COOKIE_NAME, build_ga_params
-from urlparse import parse_qs
 from google_analytics.templatetags.google_analytics_tags import google_analytics # noqa
 from google_analytics.middleware import GoogleAnalyticsMiddleware
 
+from six.moves.urllib.parse import parse_qs
 
-@pytest.mark.celery(task_always_eager=True)
+
 class GoogleAnalyticsTestCase(TestCase):
 
     def make_fake_request(self, url, headers={}):
@@ -146,6 +145,64 @@ class GoogleAnalyticsTestCase(TestCase):
         self.assertEqual(parse_qs(ga_dict.get('utm_url')).get('dt'), [
             '%D9%85%D8%A7-%D9%85%D8%AF%D9%89-%D8%AC%D8%A7%D9%87%D8%B2%D9%8A%D8'
             '%AA%D9%83-%D9%84%D9%84%D8%A5%D9%86%D8%AA%D8%B1%D9%86%D8%AA'])
+
+    @responses.activate
+    def test_build_ga_params_for_user_id(self):
+        request = self.make_fake_request('/somewhere/')
+
+        ga_dict_without_uid = build_ga_params(
+            request, 'ua-test-id', '/some/path/',)
+
+        ga_dict_with_uid = build_ga_params(
+            request, 'ua-test-id', '/some/path/', user_id='402-3a6')
+
+        self.assertEqual(
+            parse_qs(ga_dict_without_uid.get('utm_url')).get('uid'), None)
+        self.assertEqual(
+            parse_qs(ga_dict_with_uid.get('utm_url')).get('uid'), ['402-3a6'])
+
+    @responses.activate
+    @override_settings(
+        ENABLE_GA_LOGGING=True)
+    def test_ga_logging_enabled(self):
+        request = self.make_fake_request('/somewhere/')
+        ga_with_logging = build_ga_params(
+            request,
+            'ua-test-id',
+            '/some/path/',
+            user_id='402-3a6',
+        )
+        self.assertEqual(True, ga_with_logging['ga_logging_enabled'])
+
+    @responses.activate
+    @override_settings(
+        ENABLE_GA_LOGGING=False)
+    def test_ga_logging_disabled(self):
+        request = self.make_fake_request('/somewhere/')
+        ga_with_logging = build_ga_params(
+            request,
+            'ua-test-id',
+            '/some/path/',
+            user_id='402-3a6',
+        )
+
+        self.assertEqual(False, ga_with_logging['ga_logging_enabled'])
+
+    @responses.activate
+    def test_build_ga_params_for_custom_params(self):
+        request = self.make_fake_request('/somewhere/')
+
+        ga_dict_without_custom = build_ga_params(
+            request, 'ua-test-id', '/some/path/',)
+
+        ga_dict_with_custom = build_ga_params(
+            request, 'ua-test-id', '/some/path/',
+            custom_params={'key': 'value'})
+
+        self.assertEqual(
+            parse_qs(ga_dict_without_custom.get('utm_url')).get('key'), None)
+        self.assertEqual(
+            parse_qs(ga_dict_with_custom.get('utm_url')).get('key'), ['value'])
 
     @override_settings(MIDDLEWARE_CLASSES=[
         'django.contrib.sessions.middleware.SessionMiddleware',