+++ google_analytics/utils.py
@@ -1,12 +1,15 @@
 import random
 import time
-import urllib
 import uuid
+import structlog
 
 from django.conf import settings
 from django.utils.translation import get_language_from_request
 
 from google_analytics import CAMPAIGN_TRACKING_PARAMS
+
+from six import text_type
+from six.moves.urllib.parse import quote, urlencode
 
 VERSION = '1'
 COOKIE_NAME = '__utmmobile'
@@ -48,7 +51,9 @@ def set_cookie(params, response):
     return response
 
 
-def build_ga_params(request, account, path=None, event=None, referer=None, title=None):
+def build_ga_params(
+        request, account, path=None, event=None, referer=None, title=None,
+        user_id=None, custom_params={}):
     meta = request.META
     # determine the domian
     domain = meta.get('HTTP_HOST', '')
@@ -80,16 +85,23 @@ def build_ga_params(request, account, path=None, event=None, referer=None, title
         'dh': domain,
         'sr': '',
         'dr': referer,
-        'dp': urllib.quote(path.encode('utf-8')),
+        'dp': quote(path.encode('utf-8')),
         'tid': account,
         'cid': visitor_id,
         'uip': custom_uip or client_ip,
     }
 
+    # add user ID if exists
+    if user_id:
+        params.update({'uid': user_id})
+
+    # add custom parameters
+    params.update(custom_params)
+
     # add page title if supplied
     if title:
-        u_title = title.decode('utf-8') if isinstance(title, str) else title
-        params.update({'dt': urllib.quote(unicode(u_title).encode('utf-8'))})
+        u_title = title.decode('utf-8') if isinstance(title, bytes) else title
+        params.update({'dt': quote(text_type(u_title).encode('utf-8'))})
     # add event parameters if supplied
     if event:
         params.update({
@@ -115,7 +127,14 @@ def build_ga_params(request, account, path=None, event=None, referer=None, title
 
     # construct the gif hit url
     ga_url = "http://www.google-analytics.com/collect"
-    utm_url = ga_url + "?&" + urllib.urlencode(params)
+    utm_url = ga_url + "?&" + urlencode(params)
+    ga_logging_enabled = False
+    if hasattr(settings, 'ENABLE_GA_LOGGING'):
+        if settings.ENABLE_GA_LOGGING:
+            log = structlog.get_logger()
+            log.msg(utm_url, user_agent=user_agent)
+            ga_logging_enabled = True
+
     locale = get_language_from_request(request)
 
     return {'utm_url': utm_url,
@@ -125,4 +144,5 @@ def build_ga_params(request, account, path=None, event=None, referer=None, title
             'COOKIE_USER_PERSISTENCE': COOKIE_USER_PERSISTENCE,
             'COOKIE_NAME': COOKIE_NAME,
             'COOKIE_PATH': COOKIE_PATH,
+            'ga_logging_enabled': ga_logging_enabled
             }