+++ google_analytics/views.py
@@ -1,7 +1,8 @@
-import requests
 import re
 import struct
+from functools import reduce
 
+import requests
 from django.http import HttpResponse
 from django.views.decorators.cache import never_cache
 from google_analytics.utils import build_ga_params, set_cookie
@@ -14,7 +15,7 @@ GIF_DATA = reduce(lambda x, y: x + struct.pack('B', y),
                    0xff, 0x21, 0xf9, 0x04, 0x01, 0x00,
                    0x00, 0x00, 0x00, 0x2c, 0x00, 0x00,
                    0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
-                   0x00, 0x02, 0x01, 0x44, 0x00, 0x3b], '')
+                   0x00, 0x02, 0x01, 0x44, 0x00, 0x3b], b'')
 
 
 def get_ip(remote_address):